Lexicon shows reagents for apothecary/altar recipes
