+++ Xplat/src/main/java/vazkii/botania/api/recipe/PetalApothecaryRecipe.java
@@ -10,37 +10,18 @@ package vazkii.botania.api.recipe;
 
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.Container;
-import net.minecraft.world.item.crafting.Ingredient;
-import net.minecraft.world.item.crafting.Recipe;
 import net.minecraft.world.item.crafting.RecipeType;
 
 import org.jetbrains.annotations.NotNull;
 
 import vazkii.botania.api.BotaniaAPI;
 
-public interface PetalApothecaryRecipe extends Recipe<Container> {
+public interface PetalApothecaryRecipe extends RecipeWithReagent {
 	ResourceLocation TYPE_ID = new ResourceLocation(BotaniaAPI.MODID, "petal_apothecary");
-
-	/**
-	 * @return Ingredient matching the final item that needs to be thrown into the apothecary
-	 *         to perform a craft after a matching recipe is in.
-	 */
-	Ingredient getReagent();
 
 	@NotNull
 	@Override
 	default RecipeType<?> getType() {
 		return BuiltInRegistries.RECIPE_TYPE.get(TYPE_ID);
-	}
-
-	@Override
-	default boolean canCraftInDimensions(int width, int height) {
-		return false;
-	}
-
-	@Override
-	default boolean isSpecial() {
-		return true;
 	}
 }
