+++ Xplat/src/main/java/vazkii/botania/api/recipe/RunicAltarRecipe.java
@@ -10,16 +10,22 @@ package vazkii.botania.api.recipe;
 
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.Container;
-import net.minecraft.world.item.crafting.Recipe;
+import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.item.crafting.RecipeType;
 
 import org.jetbrains.annotations.NotNull;
 
 import vazkii.botania.api.BotaniaAPI;
+import vazkii.botania.common.block.BotaniaBlocks;
 
-public interface RunicAltarRecipe extends Recipe<Container> {
+public interface RunicAltarRecipe extends RecipeWithReagent {
 	ResourceLocation TYPE_ID = new ResourceLocation(BotaniaAPI.MODID, "runic_altar");
+
+	// TODO: read from recipe definition
+	@Override
+	default Ingredient getReagent() {
+		return Ingredient.of(BotaniaBlocks.livingrock);
+	}
 
 	int getManaUsage();
 
@@ -27,15 +33,5 @@ public interface RunicAltarRecipe extends Recipe<Container> {
 	@Override
 	default RecipeType<?> getType() {
 		return BuiltInRegistries.RECIPE_TYPE.get(TYPE_ID);
-	}
-
-	@Override
-	default boolean canCraftInDimensions(int width, int height) {
-		return false;
-	}
-
-	@Override
-	default boolean isSpecial() {
-		return true;
 	}
 }
